+++ modules/appeals_api/app/models/appeals_api/higher_level_review_submission.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module AppealsApi
+  class HigherLevelReviewSubmission < ApplicationRecord
+    attr_encrypted(:form_data, key: Settings.db_encryption_key, marshal: true, marshaler: JsonMarshal::Marshaller)
+    attr_encrypted(:auth_headers, key: Settings.db_encryption_key, marshal: true, marshaler: JsonMarshal::Marshaller)
+    attr_encrypted(:file_data, key: Settings.db_encryption_key, marshal: true, marshaler: JsonMarshal::Marshaller)
+
+    enum status: [ :pending, :submitted, :established, :errored ]
+  end
+end
